Rename message_parser, move Message into new file
