+++ soti/__main__.py
@@ -4,22 +4,20 @@ The main soti front-end script which initializes the terminal and listener threa
 
 import multiprocessing
 import cmd
-import json
-import os
 import serial.tools.list_ports
 import sys
-import datetime
 
 import serial.tools.list_ports_common
 
 from utils import help_strings
 from utils.constants import (
-    SAVE_DATA_DIR, SESSIONS_DIR, SESSION_FILE_FORMAT,
+    SESSIONS_DIR,
     NodeID, CmdID, COMM_INFO
 )
 
 from serial_reader import serial_reader
-from message_parser import log_messages, dict_to_yaml, Message
+from session_logger import log_messages, init_session_log, finalize_session_log
+from message import Message
 
 class ArgumentException(Exception): pass
 
@@ -162,57 +160,6 @@ class CommandLine(cmd.Cmd):
 # ----------------------------------------------------------
 # FUNCTIONS
 # ----------------------------------------------------------
-
-def init_session_log(port: str) -> str:
-    """Initializes the file which logs the session."""
-    if not os.path.exists(SAVE_DATA_DIR):
-        os.mkdir(SAVE_DATA_DIR)
-    if not os.path.exists(SESSIONS_DIR):
-        os.mkdir(SESSIONS_DIR)
-
-    session_start = datetime.datetime.now()
-
-    file_format = session_start.strftime(SESSION_FILE_FORMAT)
-    file_name = f"{file_format}.log"
-
-    date = session_start.strftime("%Y-%m-%d")
-    time = session_start.strftime("%H:%M:%S")
-
-    header_dict = {
-        "date": date,
-        "time": time,
-        "session-length": None,
-        "port": port,
-        "messages": ""
-    }
-
-    header = dict_to_yaml(header_dict, 0)
-
-    with open(SESSIONS_DIR / file_name, 'w', encoding="utf_8") as history:
-        history.write(header)
-
-    return file_name
-
-def finalize_session_log(file_name):
-    """Writes the session length to the log file."""
-    with open(SESSIONS_DIR / file_name, encoding="utf_8") as history:
-        log = history.read()
-
-    # get the datetime corresponding to the file name
-    session_start = datetime.datetime.strptime(file_name.strip(".log"), SESSION_FILE_FORMAT)
-
-    total_seconds = int((datetime.datetime.now() - session_start).total_seconds())
-    minutes, seconds = divmod(total_seconds, 60)
-    hours, minutes = divmod(minutes, 60)
-
-    # conditionally add hours and format the time
-    log = log.split("\n")
-    hours_str = f"{hours:02d}:" if hours else ""
-    log[2] = f"session-length: '{hours_str}{minutes:02d}:{seconds:02d}'"
-    log = "\n".join(log)
-
-    with open(SESSIONS_DIR / file_name, 'w', encoding="utf_8") as history:
-        history.write(log)
 
 def parse_send(args: str) -> tuple[str, str, dict, str]:
     """Parses arguments for `do_send`."""
