+++ soti/session_logger.py
@@ -1,35 +1,61 @@
 """Parses messages from the input queue."""
 
 import datetime
+import os
 import struct
 from enum import Enum
-from utils.constants import NodeID, CmdID, SESSIONS_DIR
+from utils.constants import NodeID, CmdID, SAVE_DATA_DIR, SESSIONS_DIR, SESSION_FILE_FORMAT
 
-class Message:
-    def __init__(self, msg_bytes: bytes, source: str):
-        """Initialize the message from a bytes object."""
-        self.bytes = msg_bytes
-        # serial parameters
-        self.priority = msg_bytes[0]
-        self.sender = NodeID(msg_bytes[1])
-        self.recipient = NodeID(msg_bytes[2])
-        self.cmd_id = CmdID(msg_bytes[3])
-        self.body = msg_bytes[4:]
-        # additional parameters
-        self.time = datetime.datetime.now().strftime("%T")
-        self.source = source
-        
-    def as_dict(self) -> dict:
-        """Return the message parameters as a dictionary."""
-        return {
-            "time": self.time,
-            "source": self.source,
-            "priority": self.priority,
-            "sender-id": self.sender,
-            "recipient-id": self.recipient,
-            "cmd": self.cmd_id,
-            "body": parse_msg_body(self.cmd_id, self.body)
-        }
+def init_session_log(port: str) -> str:
+    """Initializes the file which logs the session."""
+    if not os.path.exists(SAVE_DATA_DIR):
+        os.mkdir(SAVE_DATA_DIR)
+    if not os.path.exists(SESSIONS_DIR):
+        os.mkdir(SESSIONS_DIR)
+
+    session_start = datetime.datetime.now()
+
+    file_format = session_start.strftime(SESSION_FILE_FORMAT)
+    file_name = f"{file_format}.log"
+
+    date = session_start.strftime("%Y-%m-%d")
+    time = session_start.strftime("%H:%M:%S")
+
+    header_dict = {
+        "date": date,
+        "time": time,
+        "session-length": None,
+        "port": port,
+        "messages": ""
+    }
+
+    header = dict_to_yaml(header_dict, 0)
+
+    with open(SESSIONS_DIR / file_name, 'w', encoding="utf_8") as history:
+        history.write(header)
+
+    return file_name
+
+def finalize_session_log(file_name):
+    """Writes the session length to the log file."""
+    with open(SESSIONS_DIR / file_name, encoding="utf_8") as history:
+        log = history.read()
+
+    # get the datetime corresponding to the file name
+    session_start = datetime.datetime.strptime(file_name.strip(".log"), SESSION_FILE_FORMAT)
+
+    total_seconds = int((datetime.datetime.now() - session_start).total_seconds())
+    minutes, seconds = divmod(total_seconds, 60)
+    hours, minutes = divmod(minutes, 60)
+
+    # conditionally add hours and format the time
+    log = log.split("\n")
+    hours_str = f"{hours:02d}:" if hours else ""
+    log[2] = f"session-length: '{hours_str}{minutes:02d}:{seconds:02d}'"
+    log = "\n".join(log)
+
+    with open(SESSIONS_DIR / file_name, 'w', encoding="utf_8") as history:
+        history.write(log)
 
 
 def log_messages(write_msg_queue, output_file_name):
